acrn: override virDomainDef->uuid in acrnDomainDefineXMLFlags
Since ACRN only can handle specific UUIDs for the domains it supports
(these depend on ACRN scenario configuration used at compile time),
creating a domain with a new random generated UUID does not make any
sense. So, when allocating a VM (with an already known, ACRN supported
UUID), use this very ACRN UUID in the newly generated virDomainDef
object to store the configuration with. This ensures the domain can
later be started from this persistent configuration.

Signed-off-by: Helmut Buchsbaum <helmut.buchsbaum@opensource.tttech-industrial.com>