acrn_driver: introduce acrnIsRtvm
Since rtvm property lives in ACRN specific namespace data which is stored
as opaque namespaceData pointer in the domain definition, it always can
be derived from the actual domain definition. Provide an accessor to
the rtvm property via domain definition and use it wherever possible.
Furthermore use this acrnAllocateVm() which already uses the domain
definition and thus can access the rtvm property without the need to pass
it as parameter.

Signed-off-by: Helmut Buchsbaum <helmut.buchsbaum@opensource.tttech-industrial.com>