acrn-driver: fix a cpu offline bug
Currently, in acrnOfflineCpus, only one charater is used to store
CPU ID of the processor which will be offlined, if the CPU ID of a
processor is more than 9, acrn driver will fail to offline this
procesor.

In this patch, use four charaters to store CPU ID, can support
CPU ID (0 ~ 999).

Signed-off-by: Xiangyang Wu <xiangyang.wu@intel.com>